Update README.md
I think the behavior of using integers in inputs should be explicitly called out somewhere front-and-center. You clarify this in the issue queue and Stack Overflow, but it's pretty hard to find, and the casual evaluator may well dismiss this library because it "doesn't work." Maybe an example here and a caveat on the custom inputs page?